New function, astroid.extract_node, exported out from astroid.test_utils.
