+++ ChangeLog
@@ -3,6 +3,8 @@ Change log for the astroid package (used to be astng)
 
 --
 
+    * New function, astroid.extract_node, exported out from astroid.test_utils.
+
     * Fix a crash which occurred when the class of a namedtuple could not be inferred.
 
     * Add support for implicit namespace packages (PEP 420)
