+++ astroid/tests/unittest_brain.py
@@ -12,7 +12,6 @@ from astroid import builder
 from astroid import nodes
 from astroid.interpreter import objects
 from astroid.interpreter import util as interpreterutil
-from astroid import test_utils
 from astroid import util
 import astroid
 
@@ -82,7 +81,7 @@ class HashlibTest(unittest.TestCase):
 class NamedTupleTest(unittest.TestCase):
 
     def test_namedtuple_base(self):
-        klass = test_utils.extract_node("""
+        klass = astroid.extract_node("""
         from collections import namedtuple
 
         class X(namedtuple("X", ["a", "b", "c"])):
@@ -95,7 +94,7 @@ class NamedTupleTest(unittest.TestCase):
             self.assertFalse(anc.parent is None)
 
     def test_namedtuple_inference(self):
-        klass = test_utils.extract_node("""
+        klass = astroid.extract_node("""
         from collections import namedtuple
 
         name = "X"
@@ -110,7 +109,7 @@ class NamedTupleTest(unittest.TestCase):
                             set(base.instantiate_class().instance_attrs))
 
     def test_namedtuple_inference_failure(self):
-        klass = test_utils.extract_node("""
+        klass = astroid.extract_node("""
         from collections import namedtuple
 
         def foo(fields):
@@ -121,7 +120,7 @@ class NamedTupleTest(unittest.TestCase):
     def test_namedtuple_advanced_inference(self):
         # urlparse return an object of class ParseResult, which has a
         # namedtuple call and a mixin as base classes
-        result = test_utils.extract_node("""
+        result = astroid.extract_node("""
         import six
 
         result = __(six.moves.urllib.parse.urlparse('gopher://'))
@@ -135,7 +134,7 @@ class NamedTupleTest(unittest.TestCase):
         self.assertEqual(instance.name, 'ParseResult')
 
     def test_namedtuple_instance_attrs(self):
-        result = test_utils.extract_node('''
+        result = astroid.extract_node('''
         from collections import namedtuple
         namedtuple('a', 'a b c')(1, 2, 3) #@
         ''')
@@ -144,7 +143,7 @@ class NamedTupleTest(unittest.TestCase):
             self.assertEqual(attr[0].attrname, name)
 
     def test_namedtuple_uninferable_fields(self):
-        node = test_utils.extract_node('''
+        node = astroid.extract_node('''
         x = [A] * 2
         from collections import namedtuple
         l = namedtuple('a', x)
@@ -166,7 +165,7 @@ class ModuleExtenderTest(unittest.TestCase):
 class NoseBrainTest(unittest.TestCase):
 
     def test_nose_tools(self):
-        methods = test_utils.extract_node("""
+        methods = astroid.extract_node("""
         from nose.tools import assert_equal
         from nose.tools import assert_equals
         from nose.tools import assert_true
@@ -192,7 +191,7 @@ class NoseBrainTest(unittest.TestCase):
 class SixBrainTest(unittest.TestCase):
 
     def test_attribute_access(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = astroid.extract_node('''
         import six
         six.moves.http_client #@
         six.moves.urllib_parse #@
@@ -262,7 +261,7 @@ class SixBrainTest(unittest.TestCase):
             self.assertEqual(urlretrieve.qname(), 'urllib.request.urlretrieve')
 
     def test_from_imports(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = astroid.extract_node('''
         from six.moves import http_client
         http_client.HTTPSConnection #@
         ''')
@@ -285,7 +284,7 @@ class MultiprocessingBrainTest(unittest.TestCase):
         # Test that module attributes are working,
         # especially on Python 3.4+, where they are obtained
         # from a context.
-        module = test_utils.extract_node("""
+        module = astroid.extract_node("""
         import multiprocessing
         """)
         module = interpreterutil.do_import_module(module, 'multiprocessing')
@@ -296,7 +295,7 @@ class MultiprocessingBrainTest(unittest.TestCase):
             self.assertIsInstance(cpu_count, astroid.BoundMethod)
 
     def test_module_name(self):
-        module = test_utils.extract_node("""
+        module = astroid.extract_node("""
         import multiprocessing
         multiprocessing.SyncManager()
         """)
@@ -365,7 +364,7 @@ class MultiprocessingBrainTest(unittest.TestCase):
 class ThreadingBrainTest(unittest.TestCase):
 
     def test_threading(self):
-        module = test_utils.extract_node("""
+        module = astroid.extract_node("""
         import threading
         threading.Lock()
         """)
@@ -457,7 +456,7 @@ class EnumBrainTest(unittest.TestCase):
                         'IntEnum based enums should be a subtype of int')
 
     def test_enum_func_form(self):
-        instance_1, instance_2 = test_utils.extract_node('''
+        instance_1, instance_2 = astroid.extract_node('''
         from enum import Enum
         f = Enum('Audience', ['a', 'b', 'c'])
         f #@
@@ -487,7 +486,7 @@ class DateutilBrainTest(unittest.TestCase):
 class NumpyBrainTest(unittest.TestCase):
 
     def test_numpy(self):
-        node = test_utils.extract_node('''
+        node = astroid.extract_node('''
         import numpy
         numpy.ones #@
         ''')
@@ -499,7 +498,7 @@ class NumpyBrainTest(unittest.TestCase):
 class PytestBrainTest(unittest.TestCase):
 
     def test_pytest(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = astroid.extract_node('''
         import pytest
         pytest #@
         ''')
