+++ astroid/tests/unittest_helpers.py
@@ -8,6 +8,7 @@ import unittest
 import six
 from six.moves import builtins
 
+import astroid
 from astroid import builder
 from astroid import exceptions
 from astroid.interpreter import util as interpreterutil
@@ -43,12 +44,12 @@ class TestHelpers(unittest.TestCase):
             ('import sys\nsys#@', self._look_up_in_builtins(types.ModuleType.__name__)),
         ]
         for code, expected in pairs:
-            node = test_utils.extract_node(code)
+            node = astroid.extract_node(code)
             objtype = interpreterutil.object_type(node)
             self.assertIs(objtype, expected)
 
     def test_object_type_classes_and_functions(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = astroid.extract_node('''
         def generator():
             yield
 
@@ -111,7 +112,7 @@ class TestHelpers(unittest.TestCase):
 
     @test_utils.require_version(minver='3.0')
     def test_object_type_most_derived(self):
-        node = test_utils.extract_node('''
+        node = astroid.extract_node('''
         class A(type):
             def __new__(*args, **kwargs):
                  return type.__new__(*args, **kwargs)
@@ -128,14 +129,14 @@ class TestHelpers(unittest.TestCase):
         self.assertEqual(metaclass, obj_type)
 
     def test_inference_errors(self):
-        node = test_utils.extract_node('''
+        node = astroid.extract_node('''
         from unknown import Unknown
         u = Unknown #@
         ''')
         self.assertEqual(interpreterutil.object_type(node), util.Uninferable)
 
     def test_object_type_too_many_types(self):
-        node = test_utils.extract_node('''
+        node = astroid.extract_node('''
         from unknown import Unknown
         def test(x):
             if x:
@@ -147,7 +148,7 @@ class TestHelpers(unittest.TestCase):
         self.assertEqual(interpreterutil.object_type(node), util.Uninferable)
 
     def test_is_subtype(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = astroid.extract_node('''
         class int_subclass(int):
             pass
         class A(object): pass #@
@@ -173,7 +174,7 @@ class TestHelpers(unittest.TestCase):
 
     @test_utils.require_version(maxver='3.0')
     def test_is_subtype_supertype_old_style_classes(self):
-        cls_a, cls_b = test_utils.extract_node('''
+        cls_a, cls_b = astroid.extract_node('''
         class A: #@
             pass
         class B(A): #@
@@ -185,7 +186,7 @@ class TestHelpers(unittest.TestCase):
         self.assertFalse(interpreterutil.is_supertype(cls_b, cls_a))
 
     def test_is_subtype_supertype_mro_error(self):
-        cls_e, cls_f = test_utils.extract_node('''
+        cls_e, cls_f = astroid.extract_node('''
         class A(object): pass
         class B(A): pass
         class C(A): pass
@@ -200,7 +201,7 @@ class TestHelpers(unittest.TestCase):
         self.assertFalse(interpreterutil.is_supertype(cls_f, cls_e))
 
     def test_is_subtype_supertype_unknown_bases(self):
-        cls_a, cls_b = test_utils.extract_node('''
+        cls_a, cls_b = astroid.extract_node('''
         from unknown import Unknown
         class A(Unknown): pass #@
         class B(A): pass #@
@@ -211,7 +212,7 @@ class TestHelpers(unittest.TestCase):
             interpreterutil.is_supertype(cls_a, cls_b)
 
     def test_is_subtype_supertype_unrelated_classes(self):
-        cls_a, cls_b = test_utils.extract_node('''
+        cls_a, cls_b = astroid.extract_node('''
         class A(object): pass #@
         class B(object): pass #@
         ''')
@@ -221,7 +222,7 @@ class TestHelpers(unittest.TestCase):
         self.assertFalse(interpreterutil.is_supertype(cls_b, cls_a))
 
     def test_is_subtype_supertype_classes_no_type_ancestor(self):
-        cls_a = test_utils.extract_node('''
+        cls_a = astroid.extract_node('''
         class A(object): #@
             pass
         ''')
@@ -230,7 +231,7 @@ class TestHelpers(unittest.TestCase):
         self.assertFalse(interpreterutil.is_subtype(cls_a, builtin_type))
 
     def test_is_subtype_supertype_classes_metaclasses(self):
-        cls_a = test_utils.extract_node('''
+        cls_a = astroid.extract_node('''
         class A(type): #@
             pass
         ''')
@@ -240,7 +241,7 @@ class TestHelpers(unittest.TestCase):
 
     @test_utils.require_version(maxver='3.0')
     def test_old_style_class(self):
-        cls = test_utils.extract_node('''class A: pass''')
+        cls = astroid.extract_node('''class A: pass''')
         builtin_type = self._look_up_in_builtins('type')
         self.assertEqual(interpreterutil.object_type(cls), builtin_type)
