+++ astroid/tests/unittest_inference.py
@@ -12,7 +12,7 @@ import warnings
 import six
 
 from astroid import InferenceError, builder, nodes
-from astroid.builder import parse
+from astroid.builder import parse, extract_node
 from astroid import exceptions
 from astroid.inference import infer_end as inference_infer_end
 from astroid.interpreter.objects import (
@@ -326,7 +326,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             class A(A):  #@
                 pass
         '''
-        a1, a2 = test_utils.extract_node(code, __name__)
+        a1, a2 = extract_node(code, __name__)
         a2_ancestors = list(a2.ancestors())
         self.assertEqual(len(a2_ancestors), 2)
         self.assertIs(a2_ancestors[0], a1)
@@ -342,7 +342,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             class A(B):  #@
                 pass
         '''
-        a1, b, a2 = test_utils.extract_node(code, __name__)
+        a1, b, a2 = extract_node(code, __name__)
         a2_ancestors = list(a2.ancestors())
         self.assertEqual(len(a2_ancestors), 3)
         self.assertIs(a2_ancestors[0], b)
@@ -381,7 +381,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         def f():
             raise __(NotImplementedError)
         '''
-        error = test_utils.extract_node(code, __name__)
+        error = extract_node(code, __name__)
         nie = error.inferred()[0]
         self.assertIsInstance(nie, nodes.ClassDef)
         nie_ancestors = [c.name for c in nie.ancestors()]
@@ -418,7 +418,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         code = '''
             del undefined_attr
         '''
-        delete = test_utils.extract_node(code, __name__)
+        delete = extract_node(code, __name__)
         self.assertRaises(InferenceError, delete.infer)
 
     def test_del2(self):
@@ -587,7 +587,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
                 tags = list(tags)
                 __(tags).append(vid)
         '''
-        name = test_utils.extract_node(code, __name__)
+        name = extract_node(code, __name__)
         it = name.infer()
         tags = next(it)
         self.assertIsInstance(tags, nodes.List)
@@ -649,7 +649,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
                     return something
                 return __(something).conjugate()
         '''
-        func, retval = test_utils.extract_node(code, __name__)
+        func, retval = extract_node(code, __name__)
         self.assertEqual(
             [i.value for i in func.ilookup('something')],
             [1.0, 1.0j])
@@ -688,7 +688,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             A()[0] #@
             A()[-1] #@
         '''
-        ast_nodes = test_utils.extract_node(code, __name__)
+        ast_nodes = extract_node(code, __name__)
         expected = [1, 2, 3, 6, 'value', 'f', 3, 6, 42, 41]
         for node, expected_value in zip(ast_nodes, expected):
             inferred = next(node.infer())
@@ -696,7 +696,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred.value, expected_value)
 
     def test_invalid_subscripts(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class NoGetitem(object):
             pass
         class InvalidGetitem(object):
@@ -711,7 +711,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertRaises(InferenceError, next, node.infer())
         for node in ast_nodes[3:]:
             self.assertEqual(next(node.infer()), util.Uninferable)
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         [1, 2, 3][None] #@
         'lala'['bala'] #@
         ''')
@@ -719,7 +719,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertRaises(InferenceError, next, node.infer())
 
     def test_bytes_subscript(self):
-        node = test_utils.extract_node('''b'a'[0]''')
+        node = extract_node('''b'a'[0]''')
         inferred = next(node.infer())
         self.assertIsInstance(inferred, nodes.Const)
         if six.PY2:
@@ -778,7 +778,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         '''
         # XXX failing since __builtin__.help assignment has
         #     been moved into a function...
-        node = test_utils.extract_node(code, __name__)
+        node = extract_node(code, __name__)
         inferred = list(node.func.infer())
         self.assertEqual(len(inferred), 1, inferred)
         self.assertIsInstance(inferred[0], Instance)
@@ -788,7 +788,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         code = '''
             open("toto.txt")
         '''
-        node = test_utils.extract_node(code, __name__).func
+        node = extract_node(code, __name__).func
         inferred = list(node.infer())
         self.assertEqual(len(inferred), 1)
         if hasattr(sys, 'pypy_version_info'):
@@ -886,7 +886,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self._test_const_inferred(ast['b'], True)
 
     def test_unary_op_numbers(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         +1 #@
         -1 #@
         ~1 #@
@@ -900,7 +900,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     @test_utils.require_version(minver='3.5')
     def test_matmul(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class Array:
             def __matmul__(self, other):
                 return 42
@@ -947,7 +947,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self._test_const_inferred(ast['a'], 23<<1)
 
     def test_binary_op_other_type(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A:
             def __add__(self, other):
                 return other + 42
@@ -962,7 +962,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(second, util.Uninferable)
 
     def test_binary_op_other_type_using_reflected_operands(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(object):
             def __radd__(self, other):
                 return other + 42
@@ -977,7 +977,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(second.value, 43)
 
     def test_binary_op_reflected_and_not_implemented_is_type_error(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __radd__(self, other): return NotImplemented
 
@@ -1073,7 +1073,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         def f(g = lambda: None):
                 __(g()).x
 '''
-        callfuncnode = test_utils.extract_node(code)
+        callfuncnode = extract_node(code)
         inferred = list(callfuncnode.infer())
         self.assertEqual(len(inferred), 2, inferred)
         inferred.remove(util.Uninferable)
@@ -1332,7 +1332,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(len(inferred), 1, inferred)
 
     def test__new__bound_methods(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class cls(object): pass
         cls().__new__(cls) #@
         ''')
@@ -1533,7 +1533,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(node.type, 'function')
 
     def test_no_infinite_ancestor_loop(self):
-        klass = test_utils.extract_node("""
+        klass = extract_node("""
             import datetime
 
             def method(self):
@@ -1553,7 +1553,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
                     self.config = {0: self.config[0]}
                     self.config[0].test() #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
         expr = ast.func.expr
         self.assertRaises(InferenceError, next, expr.infer())
 
@@ -1573,7 +1573,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         tuple(1) #@
         tuple(1, 2) #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
 
         self.assertInferTuple(ast[0], [])
         self.assertInferTuple(ast[1], [1])
@@ -1605,7 +1605,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         frozenset(1) #@
         frozenset(1, 2) #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
 
         self.assertInferFrozenSet(ast[0], [])
         self.assertInferFrozenSet(ast[1], [1, 2])
@@ -1636,7 +1636,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         set(1) #@
         set(1, 2) #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
 
         self.assertInferSet(ast[0], [])
         self.assertInferSet(ast[1], [1, 2])
@@ -1667,7 +1667,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         list(1) #@
         list(1, 2) #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
         self.assertInferList(ast[0], [])
         self.assertInferList(ast[1], [1, 1, 2])
         self.assertInferList(ast[2], [1, 2, 3])
@@ -1682,7 +1682,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred.qname(), "{}.list".format(BUILTINS))
 
     def test_conversion_of_dict_methods(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         list({1:2, 2:3}.values()) #@
         list({1:2, 2:3}.keys()) #@
         tuple({1:2, 2:3}.values()) #@
@@ -1702,7 +1702,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         tuple(b"abc") #@
         set(b"abc") #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
         self.assertInferList(ast[0], [97, 98, 99])
         self.assertInferTuple(ast[1], [97, 98, 99])
         self.assertInferSet(ast[2], [97, 98, 99])
@@ -1732,7 +1732,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         dict([None,  None]) #@
 
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
         self.assertInferDict(ast[0], {})
         self.assertInferDict(ast[1], {'a': 1, 'b': 2, 'c': 3})
         for i in range(2, 5):
@@ -1750,7 +1750,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     @unittest.expectedFailure
     def test_dict_inference_call_context_not_propagated(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         def using_unknown_kwargs(**kwargs):
             return dict(**kwargs)
         using_unknown_kwargs(a=1, b=2) #@
@@ -1758,7 +1758,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertInferDict(ast_node, {'a': 1, 'b': 2})
 
     def test_dict_inference_kwargs(self):
-        ast_node = test_utils.extract_node('''dict(a=1, b=2, **{'c': 3})''')
+        ast_node = extract_node('''dict(a=1, b=2, **{'c': 3})''')
         self.assertInferDict(ast_node, {'a': 1, 'b': 2, 'c': 3})
 
     @test_utils.require_version('3.5')
@@ -1769,7 +1769,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             ('dict({"a":1}, b=2, **{"c":3})', {'a':1, 'b':2, 'c':3}),
         ]
         for code, expected_value in pairs:
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             self.assertInferDict(node, expected_value)
 
     def test_dict_invalid_args(self):
@@ -1779,7 +1779,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             'dict(**[])',
         ]
         for invalid in invalid_values:
-            ast_node = test_utils.extract_node(invalid)
+            ast_node = extract_node(invalid)
             inferred = next(ast_node.infer())
             self.assertIsInstance(inferred, Instance)
             self.assertEqual(inferred.qname(), "{}.dict".format(BUILTINS))
@@ -1808,7 +1808,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         ' '.find() #@
         ' '.count() #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
         self.assertInferConst(ast[0], u'')
         for i in range(1, 16):
             self.assertInferConst(ast[i], '')
@@ -1839,7 +1839,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         u' '.find() #@
         u' '.count() #@
         """
-        ast = test_utils.extract_node(code, __name__)
+        ast = extract_node(code, __name__)
         self.assertInferConst(ast[0], '')
         for i in range(1, 16):
             self.assertInferConst(ast[i], u'')
@@ -2026,11 +2026,11 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertRaises(InferenceError, next, module['no_yield'].infer())
 
     def test_unary_op_leaks_stop_iteration(self):
-        node = test_utils.extract_node('+[] #@')
+        node = extract_node('+[] #@')
         self.assertEqual(util.Uninferable, next(node.infer()))
 
     def test_unary_operands(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         import os
         def func(): pass
         from missing import missing
@@ -2082,7 +2082,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred, util.Uninferable)
 
     def test_unary_op_instance_method_not_callable(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A:
             __pos__ = (i for i in range(10))
         +A() #@
@@ -2090,7 +2090,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertRaises(InferenceError, next, ast_node.infer())
 
     def test_binary_op_type_errors(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         import collections
         1 + "a" #@
         1 - [] #@
@@ -2165,7 +2165,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(str(error), expected_value)
 
     def test_unary_type_errors(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         import collections
         ~[] #@
         ~() #@
@@ -2208,7 +2208,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     def test_unary_empty_type_errors(self):
         # These aren't supported right now
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         ~(2 and []) #@
         -(0 or {}) #@
         ''')
@@ -2233,7 +2233,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             ('frozenset((1, 2))', True),
         ]
         for code, expected in pairs:
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             inferred = next(node.infer())
             self.assertEqual(inferred.bool_value(), expected)
 
@@ -2303,7 +2303,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(compare.bool_value(), util.Uninferable)
 
     def test_bool_value_instances(self):
-        instances = test_utils.extract_node('''
+        instances = extract_node('''
         class FalseBoolInstance(object):
             def {bool}(self):
                 return False
@@ -2339,7 +2339,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred.bool_value(), expected_value)
 
     def test_infer_coercion_rules_for_floats_complex(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         1 + 1.0 #@
         1 * 1.0 #@
         2 - 1.0 #@
@@ -2355,7 +2355,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred.value, expected)
 
     def test_binop_list_with_elts(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         x = [A] * 1
         [1] + x
         ''')
@@ -2366,7 +2366,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertIs(inferred.elts[1], util.Uninferable)
 
     def test_binop_same_types(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(object):
             def __add__(self, other):
                 return 42
@@ -2382,7 +2382,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred.value, expected)
 
     def test_binop_different_types_reflected_only(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             pass
         class B(object):
@@ -2395,7 +2395,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_binop_different_types_normal_not_implemented_and_reflected(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             def __add__(self, other):
                 return NotImplemented
@@ -2409,7 +2409,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_binop_different_types_no_method_implemented(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             pass
         class B(object): pass
@@ -2419,7 +2419,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred, util.Uninferable)
 
     def test_binop_different_types_reflected_and_normal_not_implemented(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             def __add__(self, other): return NotImplemented
         class B(object):
@@ -2430,7 +2430,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred, util.Uninferable)
 
     def test_binop_subtype(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object): pass
         class B(A):
             def __add__(self, other): return other
@@ -2441,7 +2441,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_binop_subtype_implemented_in_parent(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             def __add__(self, other): return other
         class B(A): pass
@@ -2452,7 +2452,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_binop_subtype_not_implemented(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             pass
         class B(A):
@@ -2463,7 +2463,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred, util.Uninferable)
 
     def test_binop_supertype(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             pass
         class B(A):
@@ -2476,7 +2476,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_binop_supertype_rop_not_implemented(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             def __add__(self, other):
                 return other
@@ -2490,7 +2490,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'B')
 
     def test_binop_supertype_both_not_implemented(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         class A(object):
             def __add__(self): return NotImplemented
         class B(A):
@@ -2502,7 +2502,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred, util.Uninferable)
 
     def test_binop_inferrence_errors(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         from unknown import Unknown
         class A(object):
            def __add__(self, other): return NotImplemented
@@ -2517,7 +2517,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(next(node.infer()), util.Uninferable)
 
     def test_binop_ambiguity(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(object):
            def __add__(self, other):
                if isinstance(other, B):
@@ -2540,7 +2540,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(next(node.infer()), util.Uninferable)
 
     def test_bin_op_supertype_more_complicated_example(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __init__(self):
                 self.foo = 42
@@ -2560,7 +2560,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(int(inferred.value), 45)
 
     def test_aug_op_same_type_not_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
             def __add__(self, other): return NotImplemented
@@ -2569,7 +2569,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(next(ast_node.infer()), util.Uninferable)
 
     def test_aug_op_same_type_aug_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return other
         f = A()
@@ -2580,7 +2580,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_aug_op_same_type_aug_not_implemented_normal_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
             def __add__(self, other): return 42
@@ -2592,7 +2592,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.value, 42)
 
     def test_aug_op_subtype_both_not_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
             def __add__(self, other): return NotImplemented
@@ -2604,7 +2604,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(next(ast_node.infer()), util.Uninferable)
 
     def test_aug_op_subtype_aug_op_is_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return 42
         class B(A):
@@ -2617,7 +2617,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.value, 42)
 
     def test_aug_op_subtype_normal_op_is_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __add__(self, other): return 42
         class B(A):
@@ -2630,7 +2630,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.value, 42)
 
     def test_aug_different_types_no_method_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object): pass
         class B(object): pass
         f = A()
@@ -2639,7 +2639,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(next(ast_node.infer()), util.Uninferable)
 
     def test_aug_different_types_augop_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return other
         class B(object): pass
@@ -2651,7 +2651,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'B')
 
     def test_aug_different_types_aug_not_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
             def __add__(self, other): return other
@@ -2664,7 +2664,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'B')
 
     def test_aug_different_types_aug_not_implemented_rop_fallback(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
             def __add__(self, other): return NotImplemented
@@ -2678,7 +2678,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_augop_supertypes_none_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object): pass
         class B(object): pass
         a = A()
@@ -2687,7 +2687,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(next(ast_node.infer()), util.Uninferable)
 
     def test_augop_supertypes_not_implemented_returned_for_all(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
             def __add__(self, other): return NotImplemented
@@ -2699,7 +2699,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(next(ast_node.infer()), util.Uninferable)
 
     def test_augop_supertypes_augop_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return other
         class B(A): pass
@@ -2711,7 +2711,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'B')
 
     def test_augop_supertypes_reflected_binop_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
         class B(A):
@@ -2724,7 +2724,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_augop_supertypes_normal_binop_implemented(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __iadd__(self, other): return NotImplemented
             def __add__(self, other): return other
@@ -2740,7 +2740,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     @unittest.expectedFailure
     def test_string_interpolation(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         "a%d%d" % (1, 2) #@
         "a%(x)s" % {"x": 42} #@
         ''')
@@ -2751,7 +2751,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred.value, expected_value)
 
     def test_mul_list_supports__index__(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class Index(object):
             def __index__(self): return 2
         class NotIndex(object): pass
@@ -2771,7 +2771,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(inferred, util.Uninferable)
 
     def test_subscript_supports__index__(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class Index(object):
             def __index__(self): return 2
         class LambdaIndex(object):
@@ -2794,7 +2794,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertRaises(InferenceError, next, ast_nodes[2].infer())
 
     def test_special_method_masquerading_as_another(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class Info(object):
             def __add__(self, other):
                 return "lala"
@@ -2808,7 +2808,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.value, "lala")
 
     def test_unary_op_assignment(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object): pass
         def pos(self):
             return 42
@@ -2822,7 +2822,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     def _slicing_test_helper(self, pairs, cls, get_elts):
         for code, expected in pairs:
-            ast_node = test_utils.extract_node(code)
+            ast_node = extract_node(code)
             inferred = next(ast_node.infer())
             self.assertIsInstance(inferred, cls)
             self.assertEqual(get_elts(inferred), expected,
@@ -2894,11 +2894,11 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             "(1, 2, 3)[1:object]",
         ]
         for code in examples:
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             self.assertRaises(InferenceError, next, node.infer())
 
     def test_instance_slicing(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(object):
             def __getitem__(self, index):
                 return [1, 2, 3, 4, 5][index]
@@ -2917,7 +2917,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual([elt.value for elt in inferred.elts], expected)
 
     def test_instance_slicing_slices(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             def __getitem__(self, index):
                 return index
@@ -2929,7 +2929,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertIsNone(inferred.upper)
 
     def test_instance_slicing_fails(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(object):
             def __getitem__(self, index):
                 return 1[index]
@@ -2940,7 +2940,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertEqual(next(node.infer()), util.Uninferable)
 
     def test_type__new__with_metaclass(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class Metaclass(type):
             pass
         class Entity(object):
@@ -2963,7 +2963,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(a_inferred.value, 1)
 
     def test_type__new__not_enough_arguments(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         type.__new__(1) #@
         type.__new__(1, 2) #@
         type.__new__(1, 2, 3) #@
@@ -2974,7 +2974,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertIsInstance(inferred, Instance)
 
     def test_type__new__invalid_mcs_argument(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class Class(object): pass
         type.__new__(1, 2, 3, 4) #@
         type.__new__(Class, 2, 3, 4) #@
@@ -2984,7 +2984,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertIsInstance(inferred, Instance)
 
     def test_type__new__invalid_name(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class Class(type): pass
         type.__new__(Class, object, 1, 2) #@
         type.__new__(Class, 1, 1, 2) #@
@@ -2995,7 +2995,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertIsInstance(inferred, Instance)
 
     def test_type__new__invalid_bases(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         type.__new__(type, 'a', 1, 2) #@
         type.__new__(type, 'a', [], 2) #@
         type.__new__(type, 'a', {}, 2) #@
@@ -3007,7 +3007,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertIsInstance(inferred, Instance)
 
     def test_type__new__invalid_attrs(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         type.__new__(type, 'a', (), ()) #@
         type.__new__(type, 'a', (), object) #@
         type.__new__(type, 'a', (), 1) #@
@@ -3019,7 +3019,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertIsInstance(inferred, Instance)
 
     def test_type__new__metaclass_lookup(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class Metaclass(type):
             def test(cls): pass
             @classmethod
@@ -3045,7 +3045,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(attr[0].value, 42)
 
     def test_type__new__metaclass_and_ancestors_lookup(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class Book(object):
              title = 'Ubik'
         class MetaBook(type):
@@ -3061,7 +3061,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
     def test_function_metaclasses(self):
         # These are not supported right now, although
         # they will be in the future.
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         import six
 
         class BookMeta(type):
@@ -3085,7 +3085,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     def test_subscript_inference_error(self):
         # Used to raise StopIteration
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class AttributeDict(dict):
             def __getitem__(self, name):
                 return self
@@ -3097,7 +3097,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertIsNone(safe_infer(ast_node.targets[0]))
 
     def test_classmethod_inferred_by_context(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class Super(object):
            def instance(cls):
               return cls()
@@ -3116,7 +3116,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'Sub')
 
     def test_infer_call_result_invalid_dunder_call_on_instance(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A:
             __call__ = 42
         class B:
@@ -3132,7 +3132,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
             self.assertRaises(InferenceError, next, inferred.infer_call_result(node))
 
     def test_context_call_for_context_managers(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A:
             def __enter__(self):
                 return self
@@ -3161,7 +3161,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(third_c.name, 'A')
 
     def test_metaclass_subclasses_arguments_are_classes_not_instances(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(type):
             def test(cls):
                 return cls        
@@ -3177,7 +3177,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'B')
 
     def test_infer_cls_in_class_methods(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(type):
             def __call__(cls):
                 cls #@
@@ -3192,7 +3192,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     @unittest.expectedFailure
     def test_metaclass_arguments_are_classes_not_instances(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(type):
             def test(cls): return cls
         A.test() #@
@@ -3203,7 +3203,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.name, 'A')
 
     def test_delayed_attributes_without_slots(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             __slots__ = ('a', )
         a = A()
@@ -3218,7 +3218,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 
     @test_utils.require_version(maxver='3.0')
     def test_delayed_attributes_with_old_style_classes(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A:
             __slots__ = ('a', )
         a = A()
@@ -3228,7 +3228,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         next(ast_node.infer()).getattr('teta')
 
     def test_delayed_attributes_writable_property(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             @property
             def test(self):
@@ -3246,7 +3246,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(value.value, "a")
 
     def test_delayed_attributes_non_writable_property(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A(object):
             @property
             def test(self):
@@ -3263,7 +3263,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(value.value, 24)
 
     def test_lambda_as_methods(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class X:
            m = lambda self, arg: self.z + arg
            z = 24
@@ -3275,7 +3275,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
         self.assertEqual(inferred.value, 28)
 
     def test_inner_value_redefined_by_subclass(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class X(object):
             M = lambda self, arg: "a"
             x = 24
@@ -3296,7 +3296,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
     def test_inner_value_redefined_by_subclass_with_mro(self):
         # This might work, but it currently doesn't due to not being able
         # to reuse inference contexts.
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class X(object):
             M = lambda self, arg: arg + 1
             x = 24
@@ -3320,7 +3320,7 @@ class InferenceTest(resources.SysPathSetup, unittest.TestCase):
 class GetattrTest(unittest.TestCase):
 
     def test_yes_when_unknown(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         from missing import Missing
         getattr(1, Unknown) #@
         getattr(Unknown, 'a') #@
@@ -3340,7 +3340,7 @@ class GetattrTest(unittest.TestCase):
             self.assertEqual(inferred, util.Uninferable, node)
 
     def test_attrname_not_string(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         getattr(1, 1) #@
         c = int
         getattr(1, c) #@
@@ -3349,7 +3349,7 @@ class GetattrTest(unittest.TestCase):
             self.assertRaises(InferenceError, next, node.infer())
 
     def test_attribute_missing(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         getattr(1, 'ala') #@
         getattr(int, 'ala') #@
         getattr(float, 'bala') #@
@@ -3359,7 +3359,7 @@ class GetattrTest(unittest.TestCase):
             self.assertRaises(InferenceError, next, node.infer())
 
     def test_default(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         getattr(1, 'ala', None) #@
         getattr(int, 'bala', int) #@
         getattr(int, 'bala', getattr(int, 'portocala', None)) #@
@@ -3377,7 +3377,7 @@ class GetattrTest(unittest.TestCase):
         self.assertIsNone(third.value)
 
     def test_lookup(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(object):
             def test(self): pass
         class B(A):
@@ -3421,7 +3421,7 @@ class GetattrTest(unittest.TestCase):
         self.assertEqual(fifth.bound.name, 'X')
 
     def test_lambda(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         getattr(lambda x: x, 'f') #@
         ''')
         inferred = next(node.infer())
@@ -3431,7 +3431,7 @@ class GetattrTest(unittest.TestCase):
 class HasattrTest(unittest.TestCase):
 
     def test_inference_errors(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         from missing import Missing
 
         hasattr(Unknown, 'ala') #@
@@ -3444,7 +3444,7 @@ class HasattrTest(unittest.TestCase):
             self.assertEqual(inferred, util.Uninferable)
 
     def test_attribute_is_missing(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A: pass
         hasattr(int, 'ala') #@
         hasattr({}, 'bala') #@
@@ -3456,7 +3456,7 @@ class HasattrTest(unittest.TestCase):
             self.assertFalse(inferred.value)
 
     def test_attribute_is_not_missing(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A(object):
             def test(self): pass
         class B(A):
@@ -3480,7 +3480,7 @@ class HasattrTest(unittest.TestCase):
             self.assertTrue(inferred.value)
 
     def test_lambda(self):
-        node = test_utils.extract_node('''
+        node = extract_node('''
         hasattr(lambda x: x, 'f') #@
         ''')
         inferred = next(node.infer())
@@ -3508,12 +3508,12 @@ class BoolOpTest(unittest.TestCase):
             ('not (True or False) and True', False),
         ]
         for code, expected_value in expected:
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             inferred = next(node.infer())
             self.assertEqual(inferred.value, expected_value)
 
     def test_yes_when_unknown(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         from unknown import unknown, any, not_any
         0 and unknown #@
         unknown or 0 #@
@@ -3524,7 +3524,7 @@ class BoolOpTest(unittest.TestCase):
             self.assertEqual(inferred, util.Uninferable)
 
     def test_other_nodes(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def test(): pass
         test and 0 #@
         1 and test #@
@@ -3555,12 +3555,12 @@ class TestCallable(unittest.TestCase):
              callable(C1) #@''', True),
         ]
         for code, expected_value in expected:
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             inferred = next(node.infer())
             self.assertEqual(inferred.value, expected_value)
 
     def test_callable_methods(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class C:
             def test(self): pass
             @staticmethod
@@ -3592,7 +3592,7 @@ class TestCallable(unittest.TestCase):
             self.assertTrue(inferred)
 
     def test_inference_errors(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         from unknown import unknown
         callable(unknown) #@
         def test():
@@ -3604,7 +3604,7 @@ class TestCallable(unittest.TestCase):
             self.assertEqual(inferred, util.Uninferable)
 
     def test_not_callable(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         callable("") #@
         callable(1) #@
         callable(True) #@
@@ -3630,7 +3630,7 @@ class TestBool(unittest.TestCase):
             ('from unknown import Unknown; __(bool(Unknown))', util.Uninferable),
         ]
         for code, expected in pairs:
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             inferred = next(node.infer())
             if expected is util.Uninferable:
                 self.assertEqual(expected, inferred)
@@ -3638,7 +3638,7 @@ class TestBool(unittest.TestCase):
                 self.assertEqual(inferred.value, expected)
 
     def test_bool_bool_special_method(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class FalseClass:
            def {method}(self):
                return False
@@ -3672,7 +3672,7 @@ class TestBool(unittest.TestCase):
             self.assertEqual(inferred.value, expected_value)
 
     def test_bool_instance_not_callable(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class BoolInvalid(object):
            {method} = 42
         class LenInvalid(object):
@@ -3698,7 +3698,7 @@ class TestType(unittest.TestCase):
             ('type(frozenset())', 'frozenset'),
         ]
         for code, expected in pairs:
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             inferred = next(node.infer())
             self.assertIsInstance(inferred, nodes.ClassDef)
             self.assertEqual(inferred.name, expected)
@@ -3721,7 +3721,7 @@ class ArgumentsTest(unittest.TestCase):
                            (3, ), (), (3, 4, 5),
                            (), (), (4, ), (4, 5),
                            (), (3, ), (), (), (3, ), (42, )]
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def func(*args):
             return args
         func() #@
@@ -3759,7 +3759,7 @@ class ArgumentsTest(unittest.TestCase):
             (1, 2, 3),
             (1, 4, 2, 3, 5, 6, 7),
         ]
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def func(a, b, *args):
             return args
         func(1, 2, *(1, ), *(2, 3)) #@
@@ -3772,7 +3772,7 @@ class ArgumentsTest(unittest.TestCase):
 
     def test_defaults(self):
         expected_values = [42, 3, 41, 42]
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def func(a, b, c=42, *args):
             return c
         func(1, 2) #@
@@ -3788,7 +3788,7 @@ class ArgumentsTest(unittest.TestCase):
     @test_utils.require_version('3.0')
     def test_kwonly_args(self):
         expected_values = [24, 24, 42, 23, 24, 24, 54]
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def test(*, f, b): return f
         test(f=24, b=33) #@
         def test(a, *, f): return f
@@ -3813,7 +3813,7 @@ class ArgumentsTest(unittest.TestCase):
             [('a', 1)],
             [('a', 'b')],
         ]
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def test(**kwargs):
              return kwargs
         test(a=1, b=2, c=3) #@
@@ -3827,7 +3827,7 @@ class ArgumentsTest(unittest.TestCase):
             self.assertEqual(value, expected_value)
 
     def test_kwargs_and_other_named_parameters(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def test(a=42, b=24, **kwargs):
             return kwargs
         test(42, 24, c=3, d=4) #@
@@ -3849,7 +3849,7 @@ class ArgumentsTest(unittest.TestCase):
 
     def test_kwargs_access_by_name(self):
         expected_values = [42, 42, 42, 24]
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def test(**kwargs):
             return kwargs['f']
         test(f=42) #@
@@ -3873,7 +3873,7 @@ class ArgumentsTest(unittest.TestCase):
             ('d', 4),
             ('f', 42),
         ]
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         def test(**kwargs):
              return kwargs
         test(a=1, b=2, **{'c': 3}, **{'d': 4}, f=42) #@
@@ -3884,7 +3884,7 @@ class ArgumentsTest(unittest.TestCase):
         self.assertEqual(value, expected_value)
 
     def test_kwargs_are_overriden(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def test(f):
              return f
         test(f=23, **{'f': 34}) #@
@@ -3897,7 +3897,7 @@ class ArgumentsTest(unittest.TestCase):
             self.assertEqual(inferred, util.Uninferable)
 
     def test_fail_to_infer_args(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         def test(a, **kwargs): return a
         test(*missing) #@
         test(*object) #@
@@ -3940,13 +3940,13 @@ class SliceTest(unittest.TestCase):
             ('[1, 2, 3][slice(0, 3, 2)]', [1, 3]),
         ]
         for node, expected_value in ast_nodes:
-            ast_node = test_utils.extract_node("__({})".format(node))
+            ast_node = extract_node("__({})".format(node))
             inferred = next(ast_node.infer())
             self.assertIsInstance(inferred, nodes.List)
             self.assertEqual([elt.value for elt in inferred.elts], expected_value)
 
     def test_slice_inference_error(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         from unknown import unknown
         [1, 2, 3][slice(None, unknown, unknown)] #@
         [1, 2, 3][slice(None, missing, missing)] #@
@@ -3968,7 +3968,7 @@ class SliceTest(unittest.TestCase):
         ]
         for code, values in ast_nodes:
             lower, upper, step = values
-            node = test_utils.extract_node(code)
+            node = extract_node(code)
             inferred = next(node.infer())
             self.assertIsInstance(inferred, nodes.Slice)
             lower_value = next(inferred.igetattr('start'))
@@ -3983,7 +3983,7 @@ class SliceTest(unittest.TestCase):
             self.assertEqual(inferred.pytype(), '%s.slice' % BUILTINS)
 
     def test_slice_type(self):
-        ast_node = test_utils.extract_node('type(slice(None, None, None))')
+        ast_node = extract_node('type(slice(None, None, None))')
         inferred = next(ast_node.infer())
         self.assertIsInstance(inferred, nodes.ClassDef)
         self.assertEqual(inferred.name, 'slice')
@@ -3992,7 +3992,7 @@ class SliceTest(unittest.TestCase):
 class DirTest(unittest.TestCase):
 
     def test_instance_custom_ok_dir(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A:
             def __dir__(self):
                 return ['a', 'b', 'c']
@@ -4029,7 +4029,7 @@ class DirTest(unittest.TestCase):
                              sorted(expected_values))
 
     def test_modules_dir(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         import collections
         class A:
             def __dir__(self):
@@ -4044,14 +4044,14 @@ class DirTest(unittest.TestCase):
             self.assertTrue({'OrderedDict', 'deque', 'defaultdict', 'namedtuple'}.issubset(elts))
 
     def test_list_dir(self):
-        ast_node = test_utils.extract_node('dir([])')
+        ast_node = extract_node('dir([])')
         inferred = next(ast_node.infer())
         elts = {elt.value for elt in inferred.elts}
         self.assertTrue({'append', 'count', 'extend', 'index', 'insert',
                          'mro', 'pop', 'remove', 'reverse', 'sort'}.issubset(elts))
 
     def test_metaclass_dir(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         import six
         class Meta(type):
             def __dir__(self):
@@ -4078,7 +4078,7 @@ class DirTest(unittest.TestCase):
             self.assertEqual(elts, ['a', 'b'])
 
     def test_class_dir(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = extract_node('''
         class A:
             x = 42
         class B(A):
@@ -4096,7 +4096,7 @@ class DirTest(unittest.TestCase):
         self.assertTrue({'__bases__', 'x', 'y', '__dict__', '__doc__'}.issubset(elts))
 
     def test_no_args_dir(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = extract_node('''
         class A:
             pass
         class B:
