+++ astroid/tests/unittest_lookup.py
@@ -12,7 +12,6 @@ from astroid import exceptions
 from astroid.interpreter import lookup
 from astroid import nodes
 from astroid.tree import scoped_nodes
-from astroid import test_utils
 from astroid import util
 from astroid.tests import resources
 
@@ -270,7 +269,7 @@ class LookupTest(resources.SysPathSetup, unittest.TestCase):
                 def test(self):
                     pass
         '''
-        member = test_utils.extract_node(code, __name__).targets[0]
+        member = builder.extract_node(code, __name__).targets[0]
         it = member.infer()
         obj = next(it)
         self.assertIsInstance(obj, nodes.Const)
@@ -287,7 +286,7 @@ class LookupTest(resources.SysPathSetup, unittest.TestCase):
                 def funcA():
                     return 4
         '''
-        decname = test_utils.extract_node(code, __name__)
+        decname = builder.extract_node(code, __name__)
         it = decname.infer()
         obj = next(it)
         self.assertIsInstance(obj, nodes.FunctionDef)
