+++ astroid/tests/unittest_protocols.py
@@ -6,11 +6,12 @@ import contextlib
 import unittest
 
 import astroid
-from astroid.test_utils import extract_node, require_version
+from astroid.test_utils import  require_version
 from astroid import InferenceError
 from astroid import nodes
 from astroid import util
 from astroid.tree.node_classes import AssignName, Const, Name, Starred
+from astroid import extract_node
 
 
 @contextlib.contextmanager
