+++ astroid/tests/unittest_scoped_nodes.py
@@ -318,7 +318,7 @@ class FunctionNodeTest(ModuleLoader, unittest.TestCase):
         self.assertFalse(func.is_abstract(pass_is_abstract=False))
 
     def test_is_abstract_decorated(self):
-        methods = test_utils.extract_node("""
+        methods = builder.extract_node("""
             import abc
 
             class Klass(object):
@@ -540,7 +540,7 @@ class FunctionNodeTest(ModuleLoader, unittest.TestCase):
                          'classmethod')
 
     def test_igetattr(self):
-        func = test_utils.extract_node('''
+        func = builder.extract_node('''
         def test():
             pass
         ''')
@@ -576,7 +576,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         if not cls.newstyle:
             self.assertRaises(AttributeInferenceError, cls.getattr, '__mro__')
 
-        ast_list = test_utils.extract_node("[1]")
+        ast_list = builder.extract_node("[1]")
         for cls in (ast_list._proxied,
                     nodes.Const(1)._proxied):
             self.assertEqual(len(cls.getattr('__bases__')), 1)
@@ -588,7 +588,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
             self.assertEqual(len(cls.getattr('__mro__')), 1)
 
     def test__mro__attribute(self):
-        node = test_utils.extract_node('''
+        node = builder.extract_node('''
         class A(object): pass
         class B(object): pass
         class C(A, B): pass        
@@ -598,7 +598,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(mro.elts, node.mro())
 
     def test__bases__attribute(self):
-        node = test_utils.extract_node('''
+        node = builder.extract_node('''
         class A(object): pass
         class B(object): pass
         class C(A, B): pass
@@ -1037,7 +1037,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
                 self.assertEqual(meta.name, metaclass)
 
     def test_metaclass_type(self):
-        klass = test_utils.extract_node("""
+        klass = builder.extract_node("""
             def with_metaclass(meta, base=object):
                 return meta("NewBase", (base, ), {})
 
@@ -1049,7 +1049,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
             [base.name for base in klass.ancestors()])
 
     def test_no_infinite_metaclass_loop(self):
-        klass = test_utils.extract_node("""
+        klass = builder.extract_node("""
             class SSS(object):
 
                 class JJJ(object):
@@ -1071,7 +1071,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertIn('JJJ', ancestors)
 
     def test_no_infinite_metaclass_loop_with_redefine(self):
-        nodes = test_utils.extract_node("""
+        nodes = builder.extract_node("""
             import datetime
 
             class A(datetime.date): #@
@@ -1089,7 +1089,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
             self.assertEqual(None, klass.metaclass())
 
     def test_metaclass_generator_hack(self):
-        klass = test_utils.extract_node("""
+        klass = builder.extract_node("""
             import six
 
             class WithMeta(six.with_metaclass(type, object)): #@
@@ -1102,7 +1102,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
             'type', klass.metaclass().name)
 
     def test_using_six_add_metaclass(self):
-        klass = test_utils.extract_node('''
+        klass = builder.extract_node('''
         import six
         import abc
 
@@ -1116,7 +1116,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(metaclass.qname(), 'abc.ABCMeta')
 
     def test_using_invalid_six_add_metaclass_call(self):
-        klass = test_utils.extract_node('''
+        klass = builder.extract_node('''
         import six
         @six.add_metaclass()
         class Invalid(object):
@@ -1265,7 +1265,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
 
     @test_utils.require_version(maxver='3.0')
     def test_no_mro_for_old_style(self):
-        node = test_utils.extract_node("""
+        node = builder.extract_node("""
         class Old: pass""")
         with self.assertRaises(TypeError) as cm:
             node.mro()
@@ -1274,7 +1274,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
 
     @test_utils.require_version(maxver='3.0')
     def test_combined_newstyle_oldstyle_in_mro(self):
-        node = test_utils.extract_node('''
+        node = builder.extract_node('''
         class Old:
             pass
         class New(object):
@@ -1380,7 +1380,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertIsInstance(cm.exception, ResolveError)
 
     def test_generator_from_infer_call_result_parent(self):
-        func = test_utils.extract_node("""
+        func = builder.extract_node("""
         import contextlib
 
         @contextlib.contextmanager
@@ -1392,7 +1392,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(result.parent, func)
 
     def test_type_three_arguments(self):
-        classes = test_utils.extract_node("""
+        classes = builder.extract_node("""
         type('A', (object, ), {"a": 1, "b": 2, missing: 3}) #@
         """)
         first = next(classes.infer())
@@ -1409,7 +1409,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
             first.getattr("missing")
 
     def test_implicit_metaclass(self):
-        cls = test_utils.extract_node("""
+        cls = builder.extract_node("""
         class A(object):
             pass
         """)
@@ -1417,7 +1417,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(cls.implicit_metaclass(), type_cls)
 
     def test_implicit_metaclass_lookup(self):
-        cls = test_utils.extract_node('''
+        cls = builder.extract_node('''
         class A(object):
             pass
         ''')
@@ -1428,7 +1428,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
 
     def test_metaclass_lookup_using_same_class(self):
         # Check that we don't have recursive attribute access for metaclass
-        cls = test_utils.extract_node('''
+        cls = builder.extract_node('''
         class A(object): pass            
         ''')
         self.assertEqual(len(cls.getattr('mro')), 1)
@@ -1499,13 +1499,13 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
 
     @test_utils.require_version(maxver='3.0')
     def test_implicit_metaclass_is_none(self):
-        cls = test_utils.extract_node("""
+        cls = builder.extract_node("""
         class A: pass
         """)
         self.assertIsNone(cls.implicit_metaclass())
 
     def test_local_attr_invalid_mro(self):
-        cls = test_utils.extract_node("""
+        cls = builder.extract_node("""
         # A has an invalid MRO, local_attr should fallback
         # to using .ancestors.
         class A(object, object):
@@ -1554,7 +1554,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertRaises(DuplicateBasesError, module['B'].mro)
 
     def test_instance_bound_method_lambdas(self):
-        ast_nodes = test_utils.extract_node('''
+        ast_nodes = builder.extract_node('''
         class Test(object): #@
             lam = lambda self: self
             not_method = lambda xargs: xargs
@@ -1571,7 +1571,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertIsInstance(not_method, scoped_nodes.Lambda)
 
     def test_class_extra_decorators_frame_is_not_class(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = builder.extract_node('''
         def ala():
             def bala(): #@
                 func = 42
@@ -1579,7 +1579,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(ast_node.extra_decorators, [])
 
     def test_class_extra_decorators_only_callfunc_are_considered(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = builder.extract_node('''
         class Ala(object):
              def func(self): #@
                  pass
@@ -1588,7 +1588,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(ast_node.extra_decorators, [])
 
     def test_class_extra_decorators_only_assignment_names_are_considered(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = builder.extract_node('''
         class Ala(object):
              def func(self): #@
                  pass
@@ -1599,7 +1599,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(ast_node.extra_decorators, [])
 
     def test_class_extra_decorators_only_same_name_considered(self):
-        ast_node = test_utils.extract_node('''
+        ast_node = builder.extract_node('''
         class Ala(object):
              def func(self): #@
                 pass
@@ -1609,7 +1609,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(ast_node.type, 'method')
 
     def test_class_extra_decorators(self):
-        static_method, clsmethod = test_utils.extract_node('''
+        static_method, clsmethod = builder.extract_node('''
         class Ala(object):
              def static(self): #@
                  pass
@@ -1624,7 +1624,7 @@ class ClassNodeTest(ModuleLoader, unittest.TestCase):
         self.assertEqual(static_method.type, 'staticmethod')
 
     def test_extra_decorators_only_class_level_assignments(self):
-        node = test_utils.extract_node('''
+        node = builder.extract_node('''
         def _bind(arg):
             return arg.bind
 
@@ -1658,7 +1658,7 @@ class CallSiteTest(unittest.TestCase):
         return scoped_nodes.CallSite(func, call.args, keywords)
 
     def _call_site_from_code(self, code):
-        call = test_utils.extract_node(code)
+        call = builder.extract_node(code)
         return self._inferred_call_site_from_call(call)
 
     def _inferred_call_site_from_call(self, call):
@@ -1666,7 +1666,7 @@ class CallSiteTest(unittest.TestCase):
         return self._call_site_from_call(call, inferred)
 
     def _test_call_site_pair(self, code, expected_args, expected_keywords):
-        ast_node = test_utils.extract_node(code)
+        ast_node = builder.extract_node(code)
         call_site = self._call_site_from_call(ast_node, ast_node.func)
         self.assertEqual(len(call_site.positional_arguments), len(expected_args))
         self.assertEqual([arg.value for arg in call_site.positional_arguments],
@@ -1716,7 +1716,7 @@ class CallSiteTest(unittest.TestCase):
 
     def _test_call_site_valid_arguments(self, values, invalid):
         for value in values:
-            ast_node = test_utils.extract_node(value)
+            ast_node = builder.extract_node(value)
             call_site = self._call_site_from_call(ast_node, ast_node.func)
             self.assertEqual(call_site.has_invalid_arguments(), invalid)
 
@@ -1731,7 +1731,7 @@ class CallSiteTest(unittest.TestCase):
         self._test_call_site_valid_arguments(values, invalid=False)
 
     def test_duplicated_keyword_arguments(self):
-        ast_node = test_utils.extract_node('f(f=24, **{"f": 25})')
+        ast_node = builder.extract_node('f(f=24, **{"f": 25})')
         site = self._call_site_from_call(ast_node, ast_node.func)
         self.assertIn('f', site.duplicated_keywords)
